requirements.txt
asgiref==3.5.2
boto3==1.24.22
botocore==1.27.22
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.1
Markdown==3.3.7
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
